Feat : encrypted execute option & chore : encrypted code refactor
